refactor: centralize import.meta member names in ImportMetaPlugin
Export IMPORT_META_NAMES from ImportMetaPlugin as the single source of truth
and consume it in HarmonyDetectionParserPlugin, so adding a new import.meta
member updates ESM detection automatically.
