+++ lib/dependencies/HarmonyDetectionParserPlugin.js
@@ -10,27 +10,12 @@ const EnvironmentNotSupportAsyncWarning = require("../errors/EnvironmentNotSuppo
 const DynamicExports = require("./DynamicExports");
 const HarmonyCompatibilityDependency = require("./HarmonyCompatibilityDependency");
 const HarmonyExports = require("./HarmonyExports");
+const { IMPORT_META_NAMES } = require("./ImportMetaPlugin");
 
 /** @typedef {import("../Module").BuildMeta} BuildMeta */
 /** @typedef {import("../javascript/JavascriptParser")} JavascriptParser */
 
 const PLUGIN_NAME = "HarmonyDetectionParserPlugin";
-
-// `import.meta` members that mean "this is an ES module" (mirrors the surface
-// ImportMetaPlugin/NodeStuffPlugin treat as real `import.meta`, including the
-// ESM-only `import.meta.dirname`/`filename`). webpack pragmas such as
-// `import.meta.webpackHot`/`import.meta.webpackContext` and `import.meta.url`
-// inside `new URL()`/`new Worker()` are intentionally excluded: their own
-// plugins consume them first, so they keep working in CommonJS modules.
-const IMPORT_META_NAMES = [
-	"import.meta",
-	"import.meta.url",
-	"import.meta.webpack",
-	"import.meta.main",
-	"import.meta.env",
-	"import.meta.dirname",
-	"import.meta.filename"
-];
 
 module.exports = class HarmonyDetectionParserPlugin {
 	/**
