+++ lib/dependencies/ImportMetaPlugin.js
@@ -40,6 +40,21 @@ const ModuleInitFragmentDependency = require("./ModuleInitFragmentDependency");
 
 const PLUGIN_NAME = "ImportMetaPlugin";
 
+// `import.meta` member accesses that only exist in ES modules and therefore mark
+// the module as ESM. `import.meta.dirname`/`filename` are handled by
+// NodeStuffPlugin, the rest here. The webpack pragmas `import.meta.webpackHot`
+// and `import.meta.webpackContext` are intentionally NOT listed: they are
+// usable in CommonJS modules and must not force a module to ESM.
+const IMPORT_META_NAMES = [
+	"import.meta",
+	"import.meta.url",
+	"import.meta.webpack",
+	"import.meta.main",
+	"import.meta.env",
+	"import.meta.dirname",
+	"import.meta.filename"
+];
+
 /** @type {WeakMap<Compilation, { stringify: string, env: Record<string, string> }>} */
 const compilationMetaEnvMap = new WeakMap();
 
@@ -473,3 +488,4 @@ class ImportMetaPlugin {
 }
 
 module.exports = ImportMetaPlugin;
+module.exports.IMPORT_META_NAMES = IMPORT_META_NAMES;
